old-python-code/old-old-googledrive/old-google-drive/google-drive-for-linux/google-drive-for-linux/requirement.txt
google-api-python-client
oauth2client
python-magic
